Add gitlab tests
